requirements.txt
neoscore==0.1.16
fastapi==0.85.1
uvicorn==0.19.0
python-dotenv==0.21.1
showscore==0.1.4